feat: yocto: update to have an versioned build
Now it is possible to have a versioned build and the latest master.
The PREFERRED_VERSION_slint-cpp = "x.x.x" needs to be set for versioned
build by default the latest master is build.

It is also possible to add an .bbappend or copy the recipe and modify the
recipe accordingly